requirements.txt
altair==5.1.2
asn1crypto==1.5.1
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.3
boto3==1.28.72
botocore==1.31.72
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.0.0
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
filelock==3.12.4
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
html2text==2020.1.16
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
-e .
numpy==1.24.4
oscrypto==1.3.0
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
protobuf==4.24.4
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.19.0
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
s3transfer==0.7.0
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.3.1
snowflake-snowpark-python==1.9.0
sortedcontainers==2.4.0
streamlit==1.28.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
